Reduce lockfile parser drift across dependency scripts
The SBOM generator and dependency snapshot submission script used duplicate deno.lock parsing helpers. Keeping the helpers in scripts/lib/deno-lock.ts gives both scripts one parser for npm name/version splitting, package URL construction, and lock-version validation while preserving the existing script behavior.

Constraint: Public script behavior must stay compatible with the existing SBOM and dependency snapshot tests
Rejected: Keep duplicate helpers in both scripts | duplication already diverged after the snapshot edge fix
Confidence: high
Scope-risk: narrow
Reversibility: clean
Tested: deno task test:scripts
Tested: deno fmt --check --config scripts/test.deno.json scripts/build/generate-sbom.ts scripts/lib/deno-lock.ts scripts/security/submit-dependency-snapshot.ts
Tested: git diff --check origin/main...HEAD && git diff --check
Not-tested: Full repository test suite
